requirements.txt
argo-models==2.2.1a0
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
cryptography==2.8
Deprecated==1.2.6
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.21.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
idna==2.8
importlib-metadata==0.23
jsonschema==3.1.1
kfp==0.1.32.2
kfp-server-api==0.1.18.3
kubernetes==9.0.0
more-itertools==7.2.0
oauthlib==3.1.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
tabulate==0.8.3
urllib3==1.24.3
websocket-client==0.56.0
wrapt==1.11.2
zipp==0.6.0
